+++ src/main/java/com/uniminuto/clinica/repository/CitaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface CitaRepository extends JpaRepository<Cita, Integer> {
+public interface CitaRepository extends JpaRepository<Cita, Long> {
 
     List<Cita> findAllByOrderByFechaHoraDesc();
