+++ src/main/java/com/uniminuto/clinica/repository/RecetaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RecetaRepository extends JpaRepository<Receta, Integer>   {
+public interface RecetaRepository extends JpaRepository<Receta, Long>   {
     List<Receta> findAllByOrderByFechaCreacionRegistroDesc();
 
     List<Receta> findByCitaAndMedicamento(Cita cita, Medicamento medicamento);
